memxcore/requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
litellm>=1.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
mcp>=1.0.0
watchdog>=4.0.0
rank-bm25>=0.2.2
